Check for tabquery path before running unit tests that need it.
