+++ tdvt/tdvt/config_gen/test_config.py
@@ -6,6 +6,7 @@
 import glob
 import re
 from ..resources import *
+from ..tabquery_path import TabQueryPath
 
 class TestFile(object):
     """
@@ -248,30 +249,6 @@ def build_config_name(prefix, dsname):
 
 def build_tds_name(prefix, dsname):
     return prefix + dsname + '.tds'
-
-class TabQueryPath(object):
-    def __init__(self, linux_path, mac_path, windows_path):
-        self.linux_path = linux_path
-        self.mac_path = mac_path
-        self.windows_path = windows_path
-
-    @staticmethod
-    def from_array(paths):
-        if len(paths) != 3:
-            raise IndexError
-        t = TabQueryPath(paths[0], paths[1], paths[2])
-        return t
-
-    def to_array(self):
-        return [self.linux_path, self.mac_path, self.windows_path]
-
-    def get_path(self, os):
-        if os.startswith("darwin"):
-            return self.mac_path
-        elif os.startswith("linux"):
-            return self.linux_path
-        else:
-            return self.windows_path
 
 class RunTimeTestConfig(object):
     """
